Revert "fix(cli): navigate artifact on fork source links (#786)"
This reverts commit 0c571ca33586d1588a2ea156c564ca16df4f3282.
