+++ packages/artifacts/README.md
@@ -228,28 +228,6 @@ Reusable helpers include:
 - `getLatestAiSessionIdForArtifact`
 - `getEmptyAiSessionIdForArtifact`
 - `getAiSessionIdsForArtifact`
-- `switchToArtifactAiSession`
-
-Use `switchToArtifactAiSession(store, sessionId)` when navigating to an
-artifact-owned chat from history or fork provenance. It selects the owning
-artifact before switching sessions so artifact/session auto-sync does not
-immediately move the chat back to the previously selected artifact.
-
-Apps that use artifact-creating commands from AI can compose
-`createArtifactAiChatHandoffController()` with command middleware and
-`ai.onChatFinish`. The controller keeps generic artifact/session handoff
-behavior in the package:
-
-- detects AI-invoked command results with `data.artifactTargetChange`
-- restores the source artifact while the current turn is still running
-- forks the finished source chat into a target artifact chat
-- binds source and target sessions to their artifacts
-- clears inherited draft context and updates artifact run context
-
-App-specific artifact content policy stays in hooks. For example, a document
-app can provide `onBeforeTargetFork` to copy selected blocks into an empty
-target artifact without teaching the package about that artifact type.
-
 - `getAiSessionGroupsByArtifact`
 - `getRunningAiSessionCountsByArtifact`
 - `getOwningArtifactRunContextItems`
