+++ packages/artifacts/src/ai/index.ts
@@ -30,26 +30,12 @@ export {
   ArtifactAiConfigSchema,
   createArtifactAiSlice,
 } from './artifactAiSlice';
-export {
-  createArtifactAiChatHandoffController,
-  readArtifactTargetChange,
-} from './artifactAiChatHandoff';
-export {switchToArtifactAiSession} from './artifactAiNavigation';
 export type {
   ArtifactAiConfig as ArtifactAiConfigType,
   ArtifactAiSliceState,
   CreateArtifactAiSliceOptions,
   RoomStateWithArtifactAi,
 } from './artifactAiSlice';
-export type {
-  ArtifactAiChatHandoffAfterTargetForkContext,
-  ArtifactAiChatHandoffHookContext,
-  ArtifactAiChatHandoffMessageSummary,
-  ArtifactAiChatHandoffState,
-  ArtifactAiChatHandoffTargetForkContext,
-  ArtifactAiCommandMiddleware,
-  CreateArtifactAiChatHandoffControllerOptions,
-} from './artifactAiChatHandoff';
 export {
   cleanupAiSessionArtifacts,
   getAiSessionGroupsByArtifact,
